# com/west2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/note": {
            "get": {
                "description": "查看所有事项",
                "tags": [
                    "NoteController"
                ],
                "summary": "查看事项",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pageNum",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pageSize",
                        "name": "pageSIze",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "添加一条待办事项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NoteController"
                ],
                "summary": "新增待办",
                "parameters": [
                    {
                        "description": "note",
                        "name": "string",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/note/all": {
            "delete": {
                "description": "删除 所有 事项",
                "tags": [
                    "NoteController"
                ],
                "summary": "删除事项",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/note/done": {
            "get": {
                "description": "查看所有待办事项",
                "tags": [
                    "NoteController"
                ],
                "summary": "查看事项",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pageNum",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pageSize",
                        "name": "pageSIze",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "将 多条 代办事项设置为已完成",
                "tags": [
                    "NoteController"
                ],
                "summary": "完成待办",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除 所有已经完成 事项",
                "tags": [
                    "NoteController"
                ],
                "summary": "删除事项",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/note/done/{content}": {
            "get": {
                "description": "输入关键词查询事项",
                "tags": [
                    "NoteController"
                ],
                "summary": "查询事项",
                "parameters": [
                    {
                        "type": "string",
                        "description": "content",
                        "name": "content",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pageNum",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pageSize",
                        "name": "pageSIze",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/note/done/{id}": {
            "put": {
                "description": "将 一条 代办事项设置为已完成",
                "tags": [
                    "NoteController"
                ],
                "summary": "完成待办",
                "parameters": [
                    {
                        "type": "string",
                        "description": "note",
                        "name": "string",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/note/undo": {
            "put": {
                "description": "将 多条 已完成事项设置为待办",
                "tags": [
                    "NoteController"
                ],
                "summary": "设为待办",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除 所有待办 事项",
                "tags": [
                    "NoteController"
                ],
                "summary": "删除事项",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/note/undo/{id}": {
            "put": {
                "description": "将 一条 已完成事项设置为待办",
                "tags": [
                    "NoteController"
                ],
                "summary": "设为待办",
                "parameters": [
                    {
                        "type": "string",
                        "description": "note",
                        "name": "string",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/note/{id}": {
            "delete": {
                "description": "删除 一条 事项",
                "tags": [
                    "NoteController"
                ],
                "summary": "删除事项",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "string",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/test/ping": {
            "get": {
                "description": "测试",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TestController"
                ],
                "summary": "Ping!",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserController"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户vo",
                        "name": "string",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserController"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "用户vo",
                        "name": "string",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}